added additional checks to the backup script

+++ data/srv/salt/universal-backup-cron/backup-universal.sh
@@ -57,28 +57,31 @@ backup() {
         echo "Weekly backup done: to NFS folder: $DIR_TO_NFS_WEEKLY/$(basename "$FILE")"
         curl http://192.168.0.32:9950/weeklyBackupsInc
       else
-        echo "Some error, backup in $DIR_TO_NFS_WEEKLY not exists."
+        echo "ERROR: backup in $DIR_TO_BACKUP/ not exists."
       fi
     done
   else
-    echo "$DIR_TO_BACKUP is empty, Backup not create"
+    echo "ERROR: $DIR_TO_BACKUP/ is empty, Backup not create"
   fi
 }
 
 rotate_backups() {
   LAST_BACKUP="$(ls -t $DIR_TO_NFS_WEEKLY | head -1)"
-  ALL_BACKUP_WEEK="$(find $DIR_TO_NFS_WEEKLY -type f -print0 | xargs -0)"
-  cp $DIR_TO_NFS_WEEKLY/$LAST_BACKUP $DIR_TO_NFS_MONTHLY
-  if [[ -e "$DIR_TO_NFS_MONTHLY/$LAST_BACKUP" ]]; then
-    echo "Monthly backup done. Last backup: $DIR_TO_NFS_MONTHLY/$LAST_BACKUP"
-    for FILE in $ALL_BACKUP_WEEK; do
-      rm $FILE
-    done
-    curl http://192.168.0.32:9950/monthlyBackupsInc
+  if [[ ! -z "$LAST_BACKUP" ]]; then
+    ALL_BACKUP_WEEK="$(find $DIR_TO_NFS_WEEKLY -type f -print0 | xargs -0)"
+    cp $DIR_TO_NFS_WEEKLY/$LAST_BACKUP $DIR_TO_NFS_MONTHLY
+    if [[ -e "$DIR_TO_NFS_MONTHLY/$LAST_BACKUP" ]]; then
+      echo "Monthly backup done. Last backup: $DIR_TO_NFS_MONTHLY/$LAST_BACKUP"
+      for FILE in $ALL_BACKUP_WEEK; do
+        rm $FILE
+      done
+      curl http://192.168.0.32:9950/monthlyBackupsInc
+    else
+      echo "ERROR: Monthly Backup in $DIR_TO_NFS_MONTHLY/ not exists"
+    fi
   else
-    echo "Monthly Backup in $DIR_TO_NFS_MONTHLY not exists"
+    echo "ERROR: Weekly Backups in $DIR_TO_NFS_WEEKLY/ not exists" 
   fi
-  
 }
 
 # Create backups once a week (%u - day of week (1..7); 1 is Monday)
